Fix same document open multiple times
This fixes the option that the same document can be open multiple times.

A document can be opened multiple times by opening it via different paths with symlinks. If it is opened more than twice, the dialog "Document [...] has changed on disk. Should Setzer reload it now?" opens twice and cannot be closed.

After this fix only the real path is used when creating a document. Therefore the described bug cannot occur.