Fix error when set_filename() is called with None
When set_filename() in Document is called with None (which is a valid
behaviour), and calls on_filename_change() (which needs 3 arguments) through the
Observable class, the None is lost when passing through add_change_code() which
raises an error due to the missing filename argument. This commit changes all
references to on_filename_change() to give a default None value to filename so
its value is not lost, and the program can proceed without raising any error.